[Bronze V] Title: 이상한 기호, Time: 96 ms, Memory: 5484 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -1,5 +1,5 @@
 string[] str = Console.ReadLine().Split();
-int a = int.Parse(str[0]);
-int b = int.Parse(str[1]);
+long a = long.Parse(str[0]);
+long b = long.Parse(str[1]);
 
 Console.WriteLine((a + b) * (a - b));
